update starts column type

+++ backend/db/migrations/20231001042941-create-review.js
@@ -34,7 +34,7 @@ module.exports = {
           allowNull: false,
         },
         stars: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
         createdAt: {
